fixing llm QA methods + integration
